feat: user privileges,    squashed commit of the following:
commit 9a2ad60b217c6d556ad007c38a85321194bb7076
Merge: 7755bff7a70 932a8913b79
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Tue Sep 13 15:04:54 2022 +0800

    Merge branch 'feat/devenv2' of github.com:vikadata/vikadata into feat/devenv2

commit 7755bff7a703984423825796b9445a3c815bbd1d
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Tue Sep 13 15:04:48 2022 +0800

    chore: devenv under macOS

commit 932a8913b793e06d1aac3c476c657896603545ad
Merge: 14a5788b0d6 ab897cbf641
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Tue Sep 13 11:20:19 2022 +0800

    Merge branch 'feat/devenv2' of github.com:vikadata/vikadata into feat/devenv2

commit 14a5788b0d6cf9c2f49a784f37bec9cf08fb6d46
Author: Kelly Perin <kelly@vikadata.com>
Date:   Tue Sep 13 01:42:02 2022 +0000

    chore: docker-compose user to databases

commit dc9e34a514b63fcaedf63482e1b0ecd36dfe2d4d
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Tue Sep 13 03:10:54 2022 +0800

    fix: revert to grpc build for macosx

commit a20aa5c21f99cfca06b77293fa99928ba5d30bf7
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Tue Sep 13 02:41:41 2022 +0800

    chore: grpc gradle fix

commit e51cdd4202d333fbc6a2ec04736e9734eca80501
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Tue Sep 13 02:28:33 2022 +0800

    fix: protobuf  protoc compile java failed

commit db9612e236d69230d2396475fc17075bb53efa8e
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Mon Sep 12 15:14:30 2022 +0800

    chore: devenv java sdk 8 fix

commit dcc8413a031c9237910def977d1a456efffd7286
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Mon Sep 12 14:04:37 2022 +0800

    chore: devenv

commit 2f7e1bcc792526bae2ddc069fd17b78721dd4d86
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Mon Sep 12 13:51:07 2022 +0800

    chore: devenv move to apitable

commit 13c099ec67faf95fe68273a5e7b6c1c82417fd84
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Mon Sep 12 13:46:17 2022 +0800

    chore: make devenv of vikadata

commit ab897cbf64199b6b317b1c7e9b84015a11d70849
Author: Kelly Perin <kelly@vikadata.com>
Date:   Tue Sep 13 01:42:02 2022 +0000

    chore: docker-compose user to databases

commit 16e6eb07cd2d45e08764e0ba47aa331d83be0d83
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Tue Sep 13 03:10:54 2022 +0800

    fix: revert to grpc build for macosx

commit 84ce8df5fefda15109ff995ab17218c9b4f86fe0
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Tue Sep 13 02:41:41 2022 +0800

    chore: grpc gradle fix

commit 37251773be064d046e09c48c8b2ec1412c07b8a3
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Tue Sep 13 02:28:33 2022 +0800

    fix: protobuf  protoc compile java failed

commit c4d064aa9a5e19add12b77651ca1b282e588567d
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Mon Sep 12 15:14:30 2022 +0800

    chore: devenv java sdk 8 fix

commit 21a2759c65abe1ac7748e02778d06e2b280251ee
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Mon Sep 12 14:04:37 2022 +0800

    chore: devenv

commit 3c6c7565cf32d96fa0c624e89961dc801844807b
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Mon Sep 12 13:51:07 2022 +0800

    chore: devenv move to apitable

commit f4747a4af553cce3f69bb3b6fbfa743f3bf590f3
Author: Kelly Peilin Chan <kelly@vikadata.com>
Date:   Mon Sep 12 13:46:17 2022 +0800

    chore: make devenv of vikadata

+++ Makefile
@@ -12,11 +12,22 @@ GID := $(shell id -g)
 export GID
 endif
 
-
 _DATAENV := docker compose -p apitable-devenv -f docker-compose.yaml
 _DEVENV := docker compose -p apitable-devenv -f docker-compose.devenv.yaml 
-RUNNER := $(_DEVENV) run --rm --user $$UID:$$GID 
+
+OS_NAME := $(shell uname -s | tr A-Z a-z)
+ifeq ($(OS_NAME), darwin)
+    # macOS don't set user for some priviliges problem
+	RUNNER := $(_DEVENV) run --rm
+else
+    # Not found
+	RUNNER := $(_DEVENV) run --rm --user $$UID:$$GID 
+endif
 BUILDER := docker buildx bake -f docker-compose.build.yaml
+
+ttt:
+	echo $(OS_NAME)
+	echo $(RUNNER)
 
 SEMVER3 := $(shell cat .version)
 define ANNOUNCE_BODY
@@ -211,8 +222,27 @@ build: ## build all containers
 	$(BUILDER)
 
 ###### development environtments ######
+
+define DEVENV_TXT
+Which devenv do you want to start run?
+  0) UP ALL
+  1) backend-server
+  2) room-server
+  3) web-server
+  4) socket-server
+endef
+export DEVENV_TXT
+
 .PHONY: devenv
-devenv: devenv-up ## debug all devenv services with docker compose up -d
+devenv: ## debug all devenv services with docker compose up -d
+	@echo "$$DEVENV_TXT"
+	@read -p "ENTER THE NUMBER: " DEVENV_NUMBER ;\
+ 	if [ "$$DEVENV_NUMBER" = "0" ]; then make devenv-up; fi ;\
+ 	if [ "$$DEVENV_NUMBER" = "1" ]; then make devenv-backend-server; fi ;\
+ 	if [ "$$DEVENV_NUMBER" = "2" ]; then make devenv-room-server; fi ;\
+ 	if [ "$$DEVENV_NUMBER" = "3" ]; then make devenv-web-server; fi ;\
+ 	if [ "$$DEVENV_NUMBER" = "4" ]; then make devenv-socket-server; fi
+
 
 .PHONY: devenv-up
 devenv-up: ## debug all devenv services with docker compose up -d
@@ -222,7 +252,7 @@ devenv-up: ## debug all devenv services with docker compose up -d
 devenv-down: ## debug all devenv services with docker compose up -d
 	$(_DEVENV) down
 
-devenv-logs:
+devenv-logs: ## follow all devenv services logs
 	$(_DEVENV) logs -f
 devenv-ps:
 	$(_DEVENV) ps
@@ -235,28 +265,28 @@ build-backend-server:
 	$(BUILDER) backend-server
 
 .PHONY: install-backend-server
-install-backend-server: ## graldew install backend-server dependencies
+_install-backend-server: ## graldew install backend-server dependencies
 	$(RUNNER) backend-server ./gradlew build -x test
 
 .PHONY: devenv-backend-server
-devenv-backend-server: ## debug backend-server
+devenv-backend-server:
 	$(RUNNER) backend-server java -jar vikadata-service/vikadata-service-api/build/libs/vikadata-service-api.jar
 
 
 .PHONY: install-web-server
-install-web-server: ## install web-server dependencies
+_install-web-server: ## install web-server dependencies
 	$(RUNNER) web-server sh -c "yarn install"
 
 .PHONY: devenv-web-server
-devenv-web-server: ## debug web-server dependencies
-	$(RUNNER) web-server sh -c "yarn install"
+devenv-web-server:
+	$(RUNNER) web-server sh -c "yarn install && yarn sd:r"
 
 .PHONY: install-room-server
-install-room-server:
+_install-room-server:
 	$(RUNNER) room-server sh -c "yarn install && yarn build:dst:pre"
 
 .PHONY: devenv-room-server
-devenv-room-server: ## run local room-server code
+devenv-room-server: 
 	$(RUNNER) room-server yarn start:room-server
 
 
@@ -265,17 +295,20 @@ build-socket-server:
 	$(BUILDER) socket-server 
 
 .PHONY: install-socket-server
-install-socket-server:
+_install-socket-server:
 	$(RUNNER) socket-server sh -c "cd packages/socket-server/ && yarn"
 	  
 .PHONY: devenv-socket-server
-devenv-socket-server: ## run local web-server code
+devenv-socket-server:
 	$(RUNNER) socket-server sh -c "cd packages/socket-server/ && yarn run start:dev"
 
 .PHONY: install
-install: install-backend-server install-web-server install-socket-server install-room-server ## install all dependencies
+install: _install-web-server _install-backend-server _install-socket-server _install-room-server ## install all dependencies
 	echo 'Install Finished'
 
+.PHONY:
+clean: ## clean and delete git ignore and dirty files
+	git clean -fxd
 ###### buildpush ######
